requirements.txt
fastapi
fastapi-cli
uvicorn[standard]
sqlalchemy
alembic
psycopg[binary]
pydantic
pydantic-settings
python-dotenv
pyjwt
passlib[bcrypt]
bcrypt==4.0.1